+++ react/Spinner/index.jsx
@@ -4,13 +4,19 @@ import classNames from 'classnames'
 
 import styles from './styles'
 
-export const Spinner = ({ t, loadingType, middle, center, noMargin, color, size }) => {
+// === Component Options ===
+// loadingType = message type from locale [string] (default: '')
+// middle      = Component is centered both vertically & horizontally [bool] (default: false)
+// noMargin    = Component has no margin [bool] (default: false)
+// color       = Color of the spinner (available colors: blue, grey, white, red) [string] (default: 'blue')
+// size        = Size of the spinner (available sizes: tiny, small, medium, large, xlarge, xxlarge) [string] (default: 'medium')
+
+export const Spinner = ({ t, loadingType, middle, noMargin, color, size }) => {
   return (
     <div
       className={classNames(
         styles['coz-spinner'], {
           [styles['coz-spinner--middle']]: middle,
-          [styles['coz-spinner--center']]: center,
           [styles['coz-spinner--no-margin']]: noMargin,
           [styles[`coz-spinner--${color}`]]: color,
           [styles[`coz-spinner--${size}`]]: size